# echo_utilities/src/modules/logs/types/__generated__/log.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Echo API
 * Auto-generated API documentation for the Echo server
 * OpenAPI spec version: 1.0.0
 */
import type { LogCategory } from './logCategory.js'

export interface Log {
  id: string
  date: string
  groupName?: string
  fileName: string
  jobId: number
  category: LogCategory
  message: string
}